requirements.txt
asgiref==3.8.1
Django==5.1
django-admin-interface==0.28.8
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-js-asset==2.2.0
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
python-decouple==3.8
python-slugify==8.0.4
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.1
whitenoise==6.7.0
